Be

See merge request s09-webmobile1-sub2/S09P12E201!34

Former-commit-id: 3fcf95e3ba22dad9a583b0c9c2a11629de15297a

+++ backend/src/main/java/com/ssafy/api/model/MyYum.java
@@ -23,8 +23,8 @@ import lombok.ToString;
 public class MyYum {
 	@Id
 	@Column(name = "USER_ID")
-	private int userID;
+	private long userID;
 
 	@Column(name = "YUM_ID")
-	private int yumID;
+	private long yumID;
 }
